requirements.txt
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
bleach==2.0.0
boto==2.48.0
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
Django==1.9
django-cleanup==0.4.2
django-crontab==0.7.1
django-extensions==1.7.5
django-ses==0.8.4
django-smtp-ssl==1.0
django-storages==1.5.2
djangorestframework==3.4.7
entrypoints==0.2.3
enum34==1.1.6
filetype==1.0.0
flake8==3.2.1
functools32==3.2.3.post2
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.6.0
html2text==2017.10.4
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==5.4.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
kombu==4.0.2
MarkupSafe==1.0
mccabe==0.5.3
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
newrelic==2.86.0.65
notebook==5.0.0
numpy==1.13.0
olefile==0.44
packaging==16.8
pandas==0.20.2
pandocfilters==1.4.1
pathlib2==2.3.0
pdfkit==0.6.1
pdfminer==20140328
pdfrw==0.4
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycodestyle==2.2.0
pycparser==2.17
pycurl==7.43.0
pyfcm==1.2.8
pyflakes==1.3.0
Pygments==2.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
redis==2.10.5
requests==2.12.5
requests-toolbelt==0.7.1
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
tabula-py==0.9.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.21.1
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5.1
wheel==0.24.0
widgetsnbextension==2.0.0
